// src/Application/ApiCalls/generated/models/AlerteDTOOuput.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AXECIEL.GMAO.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { AlerteDTO } from './AlerteDTO';
import {
    AlerteDTOFromJSON,
    AlerteDTOFromJSONTyped,
    AlerteDTOToJSON,
    AlerteDTOToJSONTyped,
} from './AlerteDTO';

/**
 * 
 * @export
 * @interface AlerteDTOOuput
 */
export interface AlerteDTOOuput {
    /**
     * 
     * @type {string}
     * @memberof AlerteDTOOuput
     */
    message?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AlerteDTOOuput
     */
    code?: string | null;
    /**
     * 
     * @type {Array<AlerteDTO>}
     * @memberof AlerteDTOOuput
     */
    result?: Array<AlerteDTO> | null;
}

/**
 * Check if a given object implements the AlerteDTOOuput interface.
 */
export function instanceOfAlerteDTOOuput(value: object): value is AlerteDTOOuput {
    return true;
}

export function AlerteDTOOuputFromJSON(json: any): AlerteDTOOuput {
    return AlerteDTOOuputFromJSONTyped(json, false);
}

export function AlerteDTOOuputFromJSONTyped(json: any, ignoreDiscriminator: boolean): AlerteDTOOuput {
    if (json == null) {
        return json;
    }
    return {
        
        'message': json['message'] == null ? undefined : json['message'],
        'code': json['code'] == null ? undefined : json['code'],
        'result': json['result'] == null ? undefined : ((json['result'] as Array<any>).map(AlerteDTOFromJSON)),
    };
}

export function AlerteDTOOuputToJSON(json: any): AlerteDTOOuput {
    return AlerteDTOOuputToJSONTyped(json, false);
}

export function AlerteDTOOuputToJSONTyped(value?: AlerteDTOOuput | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'message': value['message'],
        'code': value['code'],
        'result': value['result'] == null ? undefined : ((value['result'] as Array<any>).map(AlerteDTOToJSON)),
    };
}